add users comments
